refactor(discord): cr round-15 (shape-drop debug log, shared places mock)
cr round-15 items:

- handleAutocomplete logs at debug when a prediction's place_id
  fails the shape check. Per-keystroke rate, so debug not warn —
  this is the upstream-shape-drift signal an operator can grep
  for when investigating "why is my dropdown empty?" after a
  hypothetical Google place_id format change.

- Extract the duplicate places mock to tests/helpers/places-mock.js.
  Three test files (qurl-file-map, commands-comprehensive,
  coverage-boost) had the same encode/decode/shape regex copy-pasted;
  the prior round-9 explicitly called out the drift risk. Single
  source of truth now lives in the helper with a JSDoc explaining
  why the `mock`-prefix on the export name matters (jest's
  babel-transform hoisting rule).

- File issue #346 for per-guild cache key dimension (round-14 #2,
  round-15 #2). Reviewer-tagged "Worth a follow-up ticket if not
  already tracked" — purely UX surprise, doesn't affect final
  destination pinning. Labeled pr-polish-followup, priority: low.

Deferred:
- _resetAutocompleteCache export discoverability (round-15 #3):
  "Pure style nit." Skip.
- INVALID_REQUEST → API_MISCONFIG reason (round-15 #4): "future
  refactor could plumb." Skip.
- autocompleteFailureBurst time-decay (round-15 #5): "Not a real
  problem." Skip.
